% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_timeseries.R
\name{as_timeseries}
\alias{as_timeseries}
\alias{as_timeseries_}
\alias{as_timeseries_.default}
\alias{as_timeseries_.data.frame}
\alias{as_timeseries_.matrix}
\alias{as_timeseries_.xts}
\alias{as_timeseries_.ts}
\alias{as_timeseries_.zooreg}
\alias{as_timeseries_.zoo}
\title{Coerce any data structure to timeSeries class}
\usage{
as_timeseries(.data, .select = NULL, .date_var = NULL,
  .silent = TRUE, ...)

as_timeseries_(.data, .select = NULL, .date_var = NULL,
  .silent = TRUE, ...)

\method{as_timeseries_}{default}(.data, .select = NULL,
  .date_var = NULL, .silent = TRUE, ...)

\method{as_timeseries_}{data.frame}(.data, .select = NULL,
  .date_var = NULL, .silent = TRUE, ...)

\method{as_timeseries_}{matrix}(.data, .select = NULL,
  .date_var = NULL, .silent = TRUE, ...)

\method{as_timeseries_}{xts}(.data, .select = NULL, .date_var = NULL,
  .silent = TRUE, ...)

\method{as_timeseries_}{ts}(.data, .select = NULL, .date_var = NULL,
  .silent = TRUE, ...)

\method{as_timeseries_}{zooreg}(.data, .select = NULL,
  .date_var = NULL, .silent = TRUE, ...)

\method{as_timeseries_}{zoo}(.data, .select = NULL, .date_var = NULL,
  .silent = TRUE, ...)
}
\arguments{
\item{.data}{The data to be coerced}

\item{.select}{Select the columns to be included in the timeSeries object. Valid only for data.frames and tibbles. If \code{NULL} (the default) all the numeric columns will be used.}

\item{.date_var}{Inform the date column. Valid for data.frames and tibbles only. If \code{NULL} the function will do it's best to find the date column.}

\item{.silent}{Should warnings messages be omitted? The default is \code{TRUE}.}

\item{...}{Addicional parameters to be passed into a \code{timeSeries} class.}
}
\value{
A timeSeries object.
}
\description{
Coerce any data structure to timeSeries class
}
\examples{

library(astimeseries)
class(EuStockMarkets)

head(as_timeseries(EuStockMarkets))
class(as_timeseries(EuStockMarkets))

}
